requirements.txt
flask
whisper
torch
ffmpeg-python
srt
indic-transliteration